\name{profile.glm}
\alias{profile.glm}
\title{
profile() for glm objects
}
\description{
This is the implementation of the profile() function for glm objects.
It profiles the scaled deviance for each parameter. 
}
\details{
Note that profile.glm requires the "model" member of the glm object.
Use  glm(\dots{}, model=T)  to include this.
}
\keyword{ellipse}
% Converted by Sd2Rd version 0.3-3.
